refactor: Move RenderOptions class to its own namespace and update references in related files
